talking_avatar_webhook/requirements.txt
Flask==3.0.2
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
python-dotenv==1.0.1
pycryptodome==3.20.0
gevent==24.2.1
gevent-websocket==0.10.1 